Add routing and refactor live events component
