+++ app/src/App.js
@@ -1,36 +1,19 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import "./App.css";
+import Live from "./Live";
 
 function App() {
-  const [events, updateEvents] = useState([]);
-
-  useEffect(() => {
-    const socket = new WebSocket("ws://localhost:8889");
-
-    socket.addEventListener("open", event =>
-      socket.send(
-        JSON.stringify({ type: "getLiveEvents", primaryMarkets: false })
-      )
-    );
-
-    socket.addEventListener("message", event => {
-      const response = JSON.parse(event.data);
-
-      // Potentially we can do more checks here to validate the data.
-      if (Array.isArray(response.data)) {
-        updateEvents(JSON.parse(event.data).data);
-      }
-    });
-  }, []);
-
   return (
     <div className="App">
       <header className="App-header">Stars Tech Test</header>
-      <ul>
-        {events.map(({ eventId, name }) => (
-          <li key={eventId}>{name}</li>
-        ))}
-      </ul>
+      <Router>
+        <Switch>
+          <Route path="/">
+            <Live />
+          </Route>
+        </Switch>
+      </Router>
     </div>
   );
 }
